[ALSA] au88x0 - Fix 64bit address of MPU401 MMIO port
Fix 64bit address of MPU401 MMIO port on au88x0 chip.

+++ sound/pci/au88x0/au88x0_mpu401.c
@@ -48,7 +48,7 @@ static int __devinit snd_vortex_midi(vortex_t * vortex)
 	struct snd_rawmidi *rmidi;
 	int temp, mode;
 	struct snd_mpu401 *mpu;
-	int port;
+	unsigned long port;
 
 #ifdef VORTEX_MPU401_LEGACY
 	/* EnableHardCodedMPU401Port() */
